rename more files
